Refactor DMA sound channel to module
